Add files via upload
This file calculates the space- and frequency dependence of the elements of the Green's function matrix, and contains routines for plotting. It is equally suitable for the calculation of fast and thermal systems. The two systems are distinguished with the values of the spectral indices Chi_p1,2 and Chi_d1,2. For a thermal system, one has to select  Chi_p1 = Chi_d1 = 1 and Chi_p2 = Chi_d2 = 0.
The input data are read direclty from my GitHub account, so no local input files are necessary, one only needs to edit the file name to be used. 
Since the code is for the simulation of 1-D systems, the data received, corresponding to a 3-D system, need to be corrected for leakaga in the neglected dimensions. This is performed in the code, one only needs to select the half-width of the system to be simulated by the 1-D slab. For instance for a cylindrical core, it is advisable to select the half-witdt of the slab to be equal to the core radius. This is made by choosing the parameter "a" in the corresponding command, which calculate the increase of the absorption cross sections to compensate for the leakaga. 
The plot routines need some manual adjustment in order that the legends fall into the proper place, since this depends on the form of the curves.